requirements.txt
google-genai
google-adk
streamlit
python-dotenv
requests
